+++ DevStyleR/C++ Programing Part 2/05. FifthLecture/04. Shop/NonAlcoholic.h
@@ -4,18 +4,18 @@
 
 class NonAlcoholic : public Drinks {
 public:
-	NonAlcoholic(std::string, double, int, std::string, double, int);
+	NonAlcoholic(std::string, double, int, std::string, double, double);
 
 	void setLiters(double);
-	void setCityTax(int);
+	void setCityTax(double);
 
 	double getLiters();
-	int    getCityTax();
+	double    getCityTax();
 
 	void   print();
 	double calculatePrice();
 
 private:
 	double _liters;
-	int _cityTax;
+	double _cityTax;
 };
